+++ src/main/java/com/trademehere/movies/MoviesApplication.java
@@ -5,14 +5,12 @@ import co.elastic.apm.attach.ElasticApmAttacher;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import com.trademehere.movies.util.EnviromentVariablesConfig;
-
 @SpringBootApplication
 public class MoviesApplication {
 
 	public static void main(String[] args) {
 		
-		ElasticApmAttacher.attach(EnviromentVariablesConfig.getElasctiAMPConfing());
+		ElasticApmAttacher.attach();
 		SpringApplication.run(MoviesApplication.class, args);
 	}