+++ src/main/java/com/trademehere/movies/dao/model/MovieDao.java
@@ -1,0 +1,28 @@
+package com.trademehere.movies.dao.model;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+import lombok.experimental.Accessors;
+
+@Getter
+@Setter
+@Accessors(chain = true)
+@NoArgsConstructor
+@ToString
+@EqualsAndHashCode(callSuper=false)
+
+@Document(collection = "movies")
+public class MovieDao {
+
+	@Id
+	private String id;
+	private String title;
+	private String movieId;
+	
+}