+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+#### Added
+
+- Data mongo-db to pom
+- DAO model Movie
+
 ## [1.0.1] - 2021-04-09
 
 ## [1.0.0] - 2021-04-09